lab 8
In this code, we can calculate a function thanks to a formula and inscribed numbers that we can specify ourselves. Added a library to open cmath several functions. Next we enter the variables that need to be entered and what to find. Then, write a small code in which you can write the value. cout << "Enter x" ; cin >> x; cout << "Enter y"; cin >> y; cout <<"Enter z" << endl; cin >> z ;
Next, we write the function itself, which is given using mathematical data. We introduce cout << "S is equal to " << S << endl;